TAO-7778 - create delivery listener, add to cleanup eligibility checking for existing delivery
