+++ scripts/install/RegisterTestCenterEvents.php
@@ -27,6 +27,7 @@ use oat\taoDeliveryRdf\model\event\DeliveryRemovedEvent;
 use oat\taoTestCenter\model\eligibility\EligiblityChanged;
 use oat\taoTestCenter\model\EligibilityService;
 use oat\taoDelivery\models\classes\execution\event\DeliveryExecutionCreated;
+use oat\taoTestCenter\model\listener\DeliveryListener;
 use oat\taoTestTaker\models\events\TestTakerRemovedEvent;
 
 /**
@@ -45,6 +46,8 @@ class RegisterTestCenterEvents extends InstallAction
         $this->registerEvent(DeliveryExecutionCreated::EVENT_NAME, [EligibilityService::SERVICE_ID, 'deliveryExecutionCreated']);
         $this->registerEvent(UserRemovedEvent::EVENT_NAME, [EligibilityService::SERVICE_ID, 'deletedTestTaker']);
         $this->registerEvent(TestTakerRemovedEvent::EVENT_NAME, [EligibilityService::SERVICE_ID, 'deletedTestTaker']);
-        $this->registerEvent(DeliveryRemovedEvent::class, [EligibilityService::SERVICE_ID, 'deleteDelivery']);
+        $this->registerEvent(DeliveryRemovedEvent::class, [DeliveryListener::SERVICE_ID, 'deleteDelivery']);
+
+        $this->getServiceManager()->register(DeliveryListener::SERVICE_ID, new DeliveryListener([]));
     }
 }
